NAS-115365: Refactoring DatasetTrivialPermissionsComponent
