+++ src/app/pages/storage/volumes/permissions/containers/dataset-trivial-permissions/dataset-trivial-permissions.component.html
@@ -9,6 +9,74 @@
   </mat-card-header>
 
   <mat-card-content>
-    <entity-form [conf]="this" class="form"></entity-form>
+    <form [formGroup]="form">
+      <div class="owner-and-access">
+        <ix-fieldset [title]="'Owner' | translate" class="owner-fields">
+          <ix-combobox
+            formControlName="user"
+            [label]="'User' | translate"
+            [tooltip]="tooltips.user | translate"
+            [provider]="userProvider"
+          ></ix-combobox>
+
+          <ix-checkbox
+            formControlName="applyUser"
+            [label]="'Apply User' | translate"
+            [tooltip]="tooltips.applyUser | translate"
+          ></ix-checkbox>
+
+          <ix-combobox
+            formControlName="group"
+            [label]="'Group' | translate"
+            [tooltip]="tooltips.group | translate"
+            [provider]="groupProvider"
+          ></ix-combobox>
+
+          <ix-checkbox
+            formControlName="applyGroup"
+            [label]="'Apply Group' | translate"
+            [tooltip]="tooltips.applyGroup | translate"
+          ></ix-checkbox>
+        </ix-fieldset>
+        <ix-fieldset [title]="'Access' | translate" class="access-fields">
+
+        </ix-fieldset>
+      </div>
+      <div class="advanced-fields">
+        <ix-checkbox
+          formControlName="recursive"
+          [label]="'Apply permissions recursively' | translate"
+          [tooltip]="tooltips.recursive | translate"
+        ></ix-checkbox>
+
+        <ix-checkbox
+          *ngIf="isRecursive$ | async"
+          formControlName="traverse"
+          [label]="'Apply permissions to child datasets' | translate"
+          [tooltip]="tooltips.traverse | translate"
+        ></ix-checkbox>
+      </div>
+
+      <div class="form-actions">
+        <button
+          mat-button
+          type="submit"
+          [disabled]="form.invalid || isLoading"
+          color="primary"
+        >{{ 'Save' | translate }}</button>
+        <button
+          mat-button
+          type="button"
+          [routerLink]="['/storage']"
+        >{{ 'Cancel' | translate }}</button>
+
+        <button
+          class="set-acl-button"
+          mat-button
+          type="button"
+          (click)="onSetAclPressed()"
+        >{{ 'Set ACL' | translate }}</button>
+      </div>
+    </form>
   </mat-card-content>
 </mat-card>
